* updated OpenNURBS to version 2012/09/14
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@7483 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ surface/include/pcl/surface/3rdparty/opennurbs/deflate.h
@@ -145,7 +145,7 @@ typedef struct internal_state {
      *   hash_shift * MIN_MATCH >= hash_bits
      */
 
-    long block_start;
+    int block_start;
     /* Window position at the beginning of the current output block. Gets
      * negative when the window is moved backwards.
      */
