+++ surface/src/3rdparty/opennurbs/adler32.c
@@ -6,7 +6,7 @@
 /* @(#) $Id$ */
 
 #define ZLIB_INTERNAL
-#include <pcl/surface/3rdparty/opennurbs/zlib.h>
+#include "pcl/surface/3rdparty/opennurbs/zlib.h"
 
 #define BASE 65521UL    /* largest prime smaller than 65536 */
 #define NMAX 5552
@@ -59,7 +59,7 @@ uLong ZEXPORT adler32(adler, buf, len)
     const Bytef *buf;
     uInt len;
 {
-    unsigned long sum2;
+    unsigned int sum2;
     unsigned n;
 
     /* split Adler-32 into component sums */
@@ -130,8 +130,8 @@ uLong ZEXPORT adler32_combine(adler1, adler2, len2)
     uLong adler2;
     z_off_t len2;
 {
-    unsigned long sum1;
-    unsigned long sum2;
+    unsigned int sum1;
+    unsigned int sum2;
     unsigned rem;
 
     /* the derivation of this formula is left as an exercise for the reader */
